refactor: progress on cleaning folder structure
