+++ Makefile
@@ -1,0 +1,5 @@
+lint-fix: 
+	gofmt -d -w .
+
+lint: 
+	gofmt -d
