+++ core/paths.go
