+++ core/snapshot_entry.go
